created sponsors and partners api + included partners in events api
