+++ backend/djangoindia/api/urls/__init__.py
@@ -1,5 +1,5 @@
 from .event import urlpatterns as event_urls
 from .communication import urlpatterns as communication_urls
-from .community_partner import urlpatterns as community_partner_urls
+from .partner_and_sponsor import urlpatterns as community_partner_urls
 
 urlpatterns = event_urls + communication_urls + community_partner_urls
