--- backend/djangoindia/api/urls/community_partner.py
@@ -1,7 +1,0 @@
-from django.urls import path
-
-from djangoindia.api.views.community_partner import CommunityPartnerAPIView
-
-urlpatterns = [
-    path("community-partners/", CommunityPartnerAPIView.as_view(), name="list_community_partners"),
-]
