+++ backend/djangoindia/api/urls/partner_and_sponsor.py
@@ -1,0 +1,7 @@
+from django.urls import path
+
+from djangoindia.api.views.partner_and_sponsor import CommunityPartnerAndSponsorAPIView
+
+urlpatterns = [
+    path("sponsors-and-partners/", CommunityPartnerAndSponsorAPIView.as_view(), name="list_community_sponsors_and_partners"),
+]
