+++ backend/djangoindia/db/models/__init__.py
@@ -1,4 +1,4 @@
 from djangoindia.db.models.event import Event, EventRegistration
 from djangoindia.db.models.communication import Subscriber, ContactUs
 from djangoindia.db.models.update import Update
-from djangoindia.db.models.community_partner import CommunityPartner
+from djangoindia.db.models.partner_and_sponsor import CommunityPartner
