+++ backend/djangoindia/db/models/update.py
@@ -1,7 +1,6 @@
 from django.db import models
 from .base import BaseModel
 from django.contrib.auth.models import User
-from django.contrib.postgres.fields import ArrayField
 
 from django.db import models
 from django.contrib.auth.models import User
